requirements.txt
pygame==2.6.1
numpy==2.1.2
Flask==3.0.3
Pillow==10.4.0
